+++ src/routes/mypage.js
@@ -1,8 +1,10 @@
 const express = require('express')
 const router = express.Router()
 
-const myAnswerRoutes = require('./mypage/myAnswer')
+const myAnswerRoutes = require('./mypage/myAnswerList')
+const myAnswerSearchRoutes = require('./mypage/myAnswerSearch')
 
-router.use('/my-answer', myAnswerRoutes)
-
+router.use('/my-answer-list', myAnswerRoutes)
+router.use('/my-answer-search', myAnswerSearchRoutes)
+// const myAnswerWriteRoutes = require('./mypage/myAnswerWrite')
 module.exports = router
